pyproject.toml
[project]
name = "kalachakra-backend"
version = "0.1.0"
description = "Backend API for KalaChakra - Vedic astrology with astronomical precision"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.8",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyswisseph>=2.10.3.2",
    "uvicorn>=0.40.0",
]
